% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swift.R
\name{get_creds}
\alias{get_creds}
\title{Get user credentials from a file}
\usage{
get_creds(fileName)
}
\arguments{
\item{fileName}{- The name of the file containing user credentials (username and password)}
}
\value{
The contents of the file
}
\description{
Extracts the user credentials from a sepcified file
and returns it.
}

