% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swift.R
\name{get_token}
\alias{get_token}
\title{Get User Token from Keystone}
\usage{
get_token(credsFileName)
}
\value{
The auth token for the user
}
\description{
Makes a call to keystone server to get a token based on the user credentials
}

