% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swift.R
\name{swift}
\alias{swift}
\title{Issue swift commands}
\usage{
swift(credsFileName, swiftCmd = NULL, container = NULL, object = NULL)
}
\arguments{
\item{credsFileName}{- Filename where user credentials are stored}
}
\description{
Issues client commands against the opoenstack swift service
}

